feat: stabilize load-from-file
